graphics/graphics-ipc: Use u32 fields in Damage
The fields should never be negative and this saves a couple of casts.

+++ graphics/driver-graphics/src/lib.rs
@@ -142,8 +142,8 @@ impl<T: GraphicsAdapter> GraphicsScheme<T> {
             &[Damage {
                 x: 0,
                 y: 0,
-                width: resource.width() as i32,
-                height: resource.height() as i32,
+                width: resource.width(),
+                height: resource.height(),
             }],
         );
     }
